Deserialize by reading directly into internal buffers if possible
