+++ bitmapcontainer.go
@@ -1,21 +1,8 @@
 package roaring
 
-import (
-	"encoding/binary"
-	"io"
-)
-
 type bitmapContainer struct {
 	cardinality int
 	bitmap      []uint64
-}
-
-func (b *bitmapContainer) readFrom(stream io.Reader) (int, error) {
-	err := binary.Read(stream, binary.LittleEndian, b.bitmap)
-	if err != nil {
-		return 0, err
-	}
-	return 8 * len(b.bitmap), nil
 }
 
 func newBitmapContainer() *bitmapContainer {
